pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-rmq"
dynamic = ["version"]
description = "RabbitMQ provider for Apache Airflow with hooks, operators, sensors, and deferrable trigger support."
readme = "readme.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "Michael Kozhin", email = "michael@kozhin.cc"},
]
keywords = ["airflow", "rabbitmq", "provider", "amqp", "pika", "aio-pika", "rmq"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "apache-airflow>=2.7.0,<3.0.0",
    "pika>=1.3.0,<2.0.0",
    "aio-pika>=9.0.0,<10.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
    "ruff>=0.1.0",
    "mypy>=1.5",
]

[project.entry-points."apache_airflow_provider"]
provider_info = "airflow_provider_rmq.__init__:get_provider_info"

[project.urls]
Homepage = "https://github.com/mkozhin/airflow-provider-rmq"
Documentation = "https://github.com/mkozhin/airflow-provider-rmq#readme"
Source = "https://github.com/mkozhin/airflow-provider-rmq"
Issues = "https://github.com/mkozhin/airflow-provider-rmq/issues"
Changelog = "https://github.com/mkozhin/airflow-provider-rmq/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["airflow_provider_rmq*"]

[tool.setuptools_scm]
version_file = "airflow_provider_rmq/_version.py"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"